feature/86cv9zk19 make a map with a center at given location
